refactor(createSchemaFile.spec.ts): move to specs dir.
