+++ src/process/output.ts
